Work in  progress

+++ actions/file_delete/actionScript.js
@@ -1,9 +1,9 @@
-const fs = require("fs");
-const path = require("path");
-const anymatch = require("anymatch");
+const fs = require('fs');
+const path = require('path');
+const anymatch = require('anymatch');
 
 //NOTE: cleanPath function prevents access to the files or folders outside files directory
-const { cleanPath } = require("./utils");
+const { cleanPath } = require('./utils');
 
 actionParameters.ExecutionResult = SUCCESS;
 try {
@@ -11,6 +11,7 @@ try {
   const directory = path.dirname(fileName);
   const mask = path.basename(fileName);
   const files = fs.readdirSync(directory).filter((fn) => anymatch(mask, fn));
+  if (files.length === 0) throw new Error(`File: ${source} does not exists`);
   files.forEach((fn) => {
     const file = path.resolve(directory, fn);
     fs.unlinkSync(file);
